# server/api-spec.json
{"openapi":"3.1.0","info":{"version":"8.9.0","title":"@fastify/swagger"},"components":{"schemas":{"Habit":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"repeatDays":{"type":"array","items":{"anyOf":[{"type":"string","enum":["Sunday"]},{"type":"string","enum":["Monday"]},{"type":"string","enum":["Tuesday"]},{"type":"string","enum":["Wednesday"]},{"type":"string","enum":["Thursday"]},{"type":"string","enum":["Friday"]},{"type":"string","enum":["Saturday"]}]}},"created":{"type":"string"},"lastUpdated":{"anyOf":[{"type":"string"},{"type":"null"}]},"userId":{"type":"number"},"category":{"anyOf":[{"type":"string"},{"type":"null"}]},"trackingStartDate":{"type":"string"}},"required":["id","name","description","repeatDays","created","lastUpdated","userId","category","trackingStartDate"]},"Completion":{"type":"object","properties":{"id":{"type":"number"},"habitId":{"type":"number"},"completionStatus":{"anyOf":[{"type":"string","enum":["complete"]},{"type":"string","enum":["incomplete"]},{"type":"string","enum":["pending"]}]},"completionStatusTimestamp":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}]},"scheduledDate":{"type":"string"}},"required":["id","habitId","completionStatus","completionStatusTimestamp","scheduledDate"]}},"securitySchemes":{"cookieAuth":{"type":"apiKey","in":"cookie","name":"sessionId"}}},"paths":{"/api/signup":{"post":{"operationId":"signup","tags":["auth"],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/api/login":{"post":{"operationId":"login","tags":["auth"],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}},"required":true},"responses":{"200":{"description":"Default Response"}}}},"/api/logout":{"post":{"operationId":"logout","tags":["auth"],"responses":{"200":{"description":"Default Response"}}}},"/api/habits":{"get":{"operationId":"getHabits","tags":["habits"],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"repeatDays":{"type":"array","items":{"anyOf":[{"type":"string","enum":["Sunday"]},{"type":"string","enum":["Monday"]},{"type":"string","enum":["Tuesday"]},{"type":"string","enum":["Wednesday"]},{"type":"string","enum":["Thursday"]},{"type":"string","enum":["Friday"]},{"type":"string","enum":["Saturday"]}]}},"created":{"type":"string"},"lastUpdated":{"anyOf":[{"type":"string"},{"type":"null"}]},"userId":{"type":"number"},"category":{"anyOf":[{"type":"string"},{"type":"null"}]},"trackingStartDate":{"type":"string"}},"required":["id","name","description","repeatDays","created","lastUpdated","userId","category","trackingStartDate"]}}}}}}},"post":{"operationId":"createHabit","tags":["habits"],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"repeatDays":{"type":"array","items":{"anyOf":[{"type":"string","enum":["Sunday"]},{"type":"string","enum":["Monday"]},{"type":"string","enum":["Tuesday"]},{"type":"string","enum":["Wednesday"]},{"type":"string","enum":["Thursday"]},{"type":"string","enum":["Friday"]},{"type":"string","enum":["Saturday"]}]}},"created":{"type":"string"},"lastUpdated":{"anyOf":[{"type":"string"},{"type":"null"}]},"userId":{"type":"number"},"category":{"anyOf":[{"type":"string"},{"type":"null"}]},"trackingStartDate":{"type":"string"}},"required":["name","repeatDays","userId","trackingStartDate"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"repeatDays":{"type":"array","items":{"anyOf":[{"type":"string","enum":["Sunday"]},{"type":"string","enum":["Monday"]},{"type":"string","enum":["Tuesday"]},{"type":"string","enum":["Wednesday"]},{"type":"string","enum":["Thursday"]},{"type":"string","enum":["Friday"]},{"type":"string","enum":["Saturday"]}]}},"created":{"type":"string"},"lastUpdated":{"anyOf":[{"type":"string"},{"type":"null"}]},"userId":{"type":"number"},"category":{"anyOf":[{"type":"string"},{"type":"null"}]},"trackingStartDate":{"type":"string"}},"required":["id","name","description","repeatDays","created","lastUpdated","userId","category","trackingStartDate"]}}}}}}},"/api/habits/{habitId}":{"delete":{"operationId":"deleteHabit","tags":["habits"],"parameters":[{"schema":{"type":"number"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/habits/completions":{"get":{"operationId":"getCompletions","tags":["habits"],"parameters":[{"schema":{"format":"date","type":"string"},"in":"query","name":"from","required":true},{"schema":{"format":"date","type":"string"},"in":"query","name":"to","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number"},"habitId":{"type":"number"},"completionStatus":{"anyOf":[{"type":"string","enum":["complete"]},{"type":"string","enum":["incomplete"]},{"type":"string","enum":["pending"]}]},"completionStatusTimestamp":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}]},"scheduledDate":{"type":"string"}},"required":["id","habitId","completionStatus","completionStatusTimestamp","scheduledDate"]}}}}}}}},"/api/habits/completions/{id}":{"patch":{"operationId":"updateCompletionStatus","tags":["habits"],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"completionStatus":{"anyOf":[{"type":"string","enum":["complete"]},{"type":"string","enum":["incomplete"]},{"type":"string","enum":["pending"]}]}},"required":["completionStatus"]}}},"required":true},"parameters":[{"schema":{"type":"number"},"in":"path","name":"id","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"number"},"habitId":{"type":"number"},"completionStatus":{"anyOf":[{"type":"string","enum":["complete"]},{"type":"string","enum":["incomplete"]},{"type":"string","enum":["pending"]}]},"completionStatusTimestamp":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}]},"scheduledDate":{"type":"string"}},"required":["id","habitId","completionStatus","completionStatusTimestamp","scheduledDate"]}}}}}}}},"security":[{"cookieAuth":[]}]}